fix: move file icon helpers out of component module
