+++ frontend/src/components/FileExplorer/index.ts
@@ -1,4 +1,5 @@
 export { FileExplorer } from './FileExplorer';
 export { TreeNode } from './TreeNode';
-export { FileIcon, getFileType, getFolderType, getFileIconColor } from './FileIcon';
+export { FileIcon } from './FileIcon';
+export { getFileType, getFolderType, getFileIconColor } from './fileIconUtils';
 export { useDirectoryTree } from './useDirectoryTree';
